add README

+++ src/public/disp_tesselate.h
@@ -1,4 +1,4 @@
-//========= Copyright © 1996-2005, Valve Corporation, All rights reserved. ============//
+//========= Copyright Valve Corporation, All rights reserved. ============//
 //
 // Purpose: 
 //
@@ -175,7 +175,7 @@ public:
 
 	
 	// Set these before calling TesselateDisplacement.
-	unsigned long *m_pActiveVerts;		// These bits control the tesselation.
+	uint32 *m_pActiveVerts;		// These bits control the tesselation.
 	const CPowerInfo *m_pPowerInfo;								// Lots of precalculated data about a displacement this size.
